+++ streetteam/apps/mediahub/urls.py
@@ -1,0 +1,4 @@
+from django.urls import path
+from .views import upload_file
+
+urlpatterns = [path("upload/", view=upload_file)]
