+++ streetteam/apps/website/urls.py
@@ -1,8 +1,4 @@
 from django.urls import path
-from .views import AccountView, IndexView, upload_file
+from .views import AccountView, IndexView
 
-urlpatterns = [
-    path("", view=IndexView.as_view()),
-    path("account/", view=AccountView.as_view()),
-    path("upload/", view=upload_file),
-]
+urlpatterns = [path("", view=IndexView.as_view()), path("account/", view=AccountView.as_view())]
