+++ streetteam/streetteam/urls.py
@@ -26,6 +26,7 @@ urlpatterns = [
     path("fubar/", admin.site.urls, name="admin"),
     path("admin/", include("admin_honeypot.urls", namespace="admin_honeypot")),
     # business functionality
+    path("", include("apps.mediahub.urls"), name="mediahub"),
     path("", include("apps.website.urls"), name="website"),
     path("sms/", include("apps.twilio_integration.urls"), name="twilio_integration"),
     # third party apps
